+++ scripts/switch.sh
@@ -15,7 +15,7 @@ function switch_proxy() {
     echo "> Port 전환"
     # nginx와 연결한 주소 생성
     # | sudo tee ~ : 앞에서 넘긴 문장을 service-url.inc에 덮어씀
-    echo "set \$service_url http://127.0.0.1:${IDLE_PORT};" | sudo tee /etc/nginx/conf.d/service-url.inc
+    echo "set \$service_url http://15.164.99.147:${IDLE_PORT};" | sudo tee /etc/nginx/conf.d/service-url.inc
 
     echo "> 엔진엑스 Reload"
     # nginx reload. restart와는 다르게 설정 값만 불러옴
