+++ src/main/java/com/f5/onepageresumebe/config/ProfileController.java
@@ -1,0 +1,31 @@
+package com.f5.onepageresumebe.config;
+
+import org.springframework.core.env.Environment;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Arrays;
+import java.util.List;
+
+@RestController
+public class ProfileController {
+
+    private final Environment env;
+
+    public ProfileController(Environment env) {
+        this.env = env;
+    }
+
+    @GetMapping("/profile")
+    public String profile() {
+        List<String> profiles = Arrays.asList(env.getActiveProfiles());
+        List<String> realProfiles = Arrays.asList("set1", "set2");
+        String defaultProfile = profiles.isEmpty()? "default" : profiles.get(0);
+
+        // set, set1, set2 중 하나라도 있으면 그 값 반환
+        return profiles.stream()
+                .filter(realProfiles::contains)
+                .findAny()
+                .orElse(defaultProfile);
+    }
+}
