--- src/main/java/com/f5/onepageresumebe/config/TestController.java
@@ -1,29 +1,0 @@
-package com.f5.onepageresumebe.config;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.env.Environment;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Arrays;
-import java.util.List;
-
-@RequiredArgsConstructor
-@RestController
-public class TestController {
-
-    private final Environment env;
-
-    @GetMapping("/profile")
-    public String profile() {
-        List<String> profile = Arrays.asList(env.getActiveProfiles());
-        List<String> realProfiles = Arrays.asList("real1", "real2");
-        String defaultProfile = profile.isEmpty() ? "default" : profile.get(0);
-
-        return profile.stream()
-                .filter(realProfiles::contains)///
-                .findAny()
-                .orElse(defaultProfile);
-    }
-}
